+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Feat: add `options.autoSticky/autoGlobal` for `fromTo`.
   - Feat: export `Action.simple/match`. Add `options.autoSticky/rejectCaret` for `Action.match`.
   - Feat: add typed error.
+  - Feat: add `ILexer.takeUntil`. [#16](https://github.com/DiscreteTom/retsac/issues/16)
 - Optimize performance.
   - [#6](https://github.com/DiscreteTom/retsac/issues/6)
   - Lexer will skip unexpected actions by checking expected text.